Simplify clock steering (#943)
Simplify clock steering

+++ src/manage.c
@@ -1181,14 +1181,14 @@ static chan_meas_flags_t compute_meas_flags(u32 flags,
  * not been yet available and feeds it back to tracker.
  *
  * \param[in]  i      Tracking channel number.
- * \param[in]  ref_tc Reference time [ticks]
+ * \param[in]  ref_tc Reference time [float ticks]
  * \param[out] meas   Container for measurement data.
  * \param[out] ephe   Container for ephemeris
  *
  * \return Flags
  */
 u32 get_tracking_channel_meas(u8 i,
-                              u64 ref_tc,
+                              double ref_tc,
                               channel_measurement_t *meas,
                               ephemeris_t *ephe) {
   u32 flags = 0;                          /* Result */
